SAE-340 Fix expiration for high TTL. Change expiration logging
High TTL (cacheTime) was calculated incorrectly due to integer overrun.
Fixed by explicit casting to long.

+++ src/main/java/com/trivago/triava/tcache/util/SecondsOrMillis.java
@@ -54,6 +54,6 @@ public class SecondsOrMillis
 			return internal;
 		else if (internal == Integer.MIN_VALUE)
 			return Long.MAX_VALUE;
-		return -internal * 1000;
+		return ((long)-internal) * 1000L;
 	}
 }
